moved 'ActionItem' to abstract
